+++ src/api/v1/index.ts
@@ -1,5 +1,6 @@
-import { Router } from 'express';
-import { emailRoutes } from './email/email.route';
+import { Router } from "express";
+
+import { emailRoutes } from "./email/email.route";
 
 export const routes = Router();