+++ src/emails/components/Button.tsx
@@ -1,8 +1,8 @@
-import { Button as EmailButton } from '@react-email/components';
+import { Button as EmailButton } from "@react-email/components";
 
 const Button = ({ children }: { children: React.ReactNode }) => {
   return (
-    <EmailButton className="bg-indigo-600 rounded font-semibold text-white text-base no-underline text-center block py-3 px-6">
+    <EmailButton className="block rounded bg-indigo-600 px-6 py-3 text-center text-base font-semibold text-white no-underline">
       {children}
     </EmailButton>
   );