+++ src/emails/components/Header.tsx
@@ -1,21 +1,22 @@
-import React from 'react';
-import { Section, Row, Column, Img, Text } from '@react-email/components';
+import React from "react";
+
+import { Column, Img, Row, Section, Text } from "@react-email/components";
 
 export const Header: React.FC = () => {
   return (
-    <Section className="bg-black rounded-t-lg p-6">
-      <Row className="flex justify-between items-center">
+    <Section className="rounded-t-lg bg-black p-6">
+      <Row className="flex items-center justify-between">
         <Column>
           <Img
             src="https://avatars.githubusercontent.com/u/123801647?s=200&v=4"
             width="40"
             height="40"
             alt="Ansopedia Logo"
-            className="inline-block align-middle mr-2"
+            className="mr-2 inline-block align-middle"
           />
         </Column>
         <Column>
-          <Text className="text-white text-2xl font-bold font-sans m-0 uppercase tracking-wider inline-block align-middle">
+          <Text className="m-0 inline-block align-middle font-sans text-2xl font-bold uppercase tracking-wider text-white">
             Ansopedia
           </Text>
         </Column>