+++ src/emails/templates/email-verification-otp.tsx
@@ -1,14 +1,25 @@
-import React from 'react';
-import { withEmailWrapper } from '../wrappers/withEmailWrapper';
-import { Section } from '@react-email/components';
-import { Typography } from '../components/ui/typography';
+import React from "react";
+
+import { Section } from "@react-email/components";
+
+import { Typography } from "../components/ui/typography";
+import { withEmailWrapper } from "../wrappers/withEmailWrapper";
 
 interface EmailVerificationOTPProps {
   otp: string;
   recipientName: string;
+  otpTTL: number; // (TTL = Time To Live)
 }
 
-const EmailVerificationOTPContent: React.FC<EmailVerificationOTPProps> = ({ otp, recipientName }) => {
+const defaultProps: EmailVerificationOTPProps = {
+  otp: "123456",
+  recipientName: "User",
+  otpTTL: 5,
+};
+
+const EmailVerificationOTPContent: React.FC<EmailVerificationOTPProps> = (props) => {
+  const { otp, recipientName, otpTTL } = { ...defaultProps, ...props };
+
   return (
     <>
       <Typography variant="h4" className="mb-6">
@@ -18,22 +29,21 @@ const EmailVerificationOTPContent: React.FC<EmailVerificationOTPProps> = ({ otp,
         Thank you for signing up with Ansopedia. To complete your registration, please use the following One-Time
         Password (OTP):
       </Typography>
-      <Typography variant="h4">This is your OTP: {otp}</Typography>
-      <Section className="bg-gray-100 rounded p-6 text-center my-6">
-        <Typography variant="h1" className="m-0 text-black">
+      <Section className="my-6 rounded bg-gray-100 p-6 text-center">
+        <Typography variant="h1" className="letter-spacing-2 m-0 text-black">
           {otp}
         </Typography>
       </Section>
       <Typography variant="p" className="mb-4">
-        This OTP is valid for 10 minutes. If you didn't request this verification, please ignore this email.
+        This OTP is valid for {otpTTL} minutes. If you didn't request this verification, please ignore this email.
       </Typography>
     </>
   );
 };
 
 export const EmailVerificationOTP = withEmailWrapper(EmailVerificationOTPContent, {
-  wrapperType: 'default',
-  previewText: 'Your Ansopedia Email Verification OTP',
+  wrapperType: "default",
+  previewText: "Your Ansopedia Email Verification OTP",
 });
 
 export default EmailVerificationOTP;