+++ src/emails/templates/forget-password-otp.tsx
@@ -1,14 +1,25 @@
-import React from 'react';
-import { withEmailWrapper } from '../wrappers/withEmailWrapper';
-import { Section } from '@react-email/components';
-import { Typography } from '../components/ui/typography';
+import React from "react";
+
+import { Section } from "@react-email/components";
+
+import { Typography } from "../components/ui/typography";
+import { withEmailWrapper } from "../wrappers/withEmailWrapper";
 
 interface ForgetPasswordOTPProps {
   otp: string;
   recipientName: string;
+  otpTTL: number; // (TTL = Time To Live)
 }
 
-const ForgetPasswordOTPContent: React.FC<ForgetPasswordOTPProps> = ({ otp, recipientName }) => {
+const defaultProps: ForgetPasswordOTPProps = {
+  otp: "123456",
+  recipientName: "User",
+  otpTTL: 5,
+};
+
+const ForgetPasswordOTPContent: React.FC<ForgetPasswordOTPProps> = (props) => {
+  const { otp, recipientName, otpTTL } = { ...defaultProps, ...props };
+
   return (
     <>
       <Typography variant="h4" className="mb-6">
@@ -19,22 +30,22 @@ const ForgetPasswordOTPContent: React.FC<ForgetPasswordOTPProps> = ({ otp, recip
         Password (OTP) to verify your identity:
       </Typography>
       <Typography variant="h4">This is your OTP: {otp}</Typography>
-      <Section className="bg-gray-100 rounded p-6 text-center my-6">
+      <Section className="my-6 rounded bg-gray-100 p-6 text-center">
         <Typography variant="h1" className="m-0 text-black">
           {otp}
         </Typography>
       </Section>
       <Typography variant="p" className="mb-4">
-        This OTP is valid for 10 minutes. If you didn't request this password reset, please ignore this email or contact
-        support immediately.
+        This OTP is valid for {otpTTL} minutes. If you didn't request this password reset, please ignore this email or
+        contact support immediately.
       </Typography>
     </>
   );
 };
 
 export const ForgetPasswordOTP = withEmailWrapper(ForgetPasswordOTPContent, {
-  wrapperType: 'default',
-  previewText: 'Your Ansopedia Password Reset OTP',
+  wrapperType: "default",
+  previewText: "Your Ansopedia Password Reset OTP",
 });
 
 export default ForgetPasswordOTP;