+++ src/utils/sendEmail.util.ts
@@ -1,7 +1,11 @@
-import nodemailer from 'nodemailer';
-import { nodemailerConfig } from '@/config';
-import Mail from 'nodemailer/lib/mailer';
-import { logger } from './Logger';
+import nodemailer from "nodemailer";
+import Mail from "nodemailer/lib/mailer";
+
+import { nodemailerConfig } from "@/config";
+
+import { logger } from "./Logger";
+
+const senderEmail = `"Ansopedia" <${nodemailerConfig.auth?.user}>`;
 
 interface EmailOptions extends Mail.Options {
   to: string;
@@ -15,11 +19,11 @@ export const sendEmail = async (options: EmailOptions) => {
     const transporter = nodemailer.createTransport(nodemailerConfig);
 
     await transporter.sendMail({
-      from: nodemailerConfig.auth?.user,
+      from: senderEmail,
       ...options,
     });
   } catch (error) {
-    logger.error('Error sending email:', error);
+    logger.error("Error sending email:", error);
     throw error;
   }
 };